requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
black==21.6b0
click==8.0.1
dj-database-url==0.5.0
Django==3.2.5
django-db==0.0.7
django-extensions==3.1.3
django-on-heroku==1.1.2
djangorestframework==3.12.4
fancycompleter==0.9.1
flake8==3.9.2
freezegun==1.1.0
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
pdbpp==0.10.3
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pyrepl==0.9.0
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
whitenoise==5.3.0
wmctrl==0.4
